separate tooltip implementation code from conf.py
